Changed worker class to gevent, added GET handler for users (for testing purposes only)
