+++ requirements.txt
@@ -1,4 +1,10 @@
 falcon==1.4.1
+greenlet==0.4.15
 gunicorn==19.9.0
+mccabe==0.6.1
+psycopg2==2.7.5
+pycodestyle==2.3.1
+pyflakes==1.6.0
 python-mimeparse==1.6.0
 six==1.11.0
+gevent==1.3.6
